初始构建中 开始构建Jump

+++ app/src/main/java/com/gjn/gamequery/net/JumpUrl.java
@@ -24,6 +24,6 @@ public interface JumpUrl {
     Observable<JumpListData> getlist(@Query("name") String name);
 
     @GET("api/getmatch")
-    Observable<JumpMatchData> getmatch(@Query("id") int id);
+    Observable<JumpMatchData> getmatch(@Query("id") long id);
 
 }
